+++ MMM-NHL.js
@@ -45,9 +45,11 @@ Module.register('MMM-NHL', {
         '3rd': '3RD_PERIOD',
         OT: 'OVER_TIME',
         SO: 'SHOOTOUT',
+        SHOOTOUT: 'SHOOTOUT',
         FINAL: 'FINAL',
         'FINAL OT': 'FINAL_OVERTIME',
-        'FINAL SO': 'FINAL_SHOOTOUT'
+        'FINAL SO': 'FINAL_SHOOTOUT',
+        PPD: 'PPD'
     },
 
     /**
@@ -58,6 +60,12 @@ Module.register('MMM-NHL', {
      * @member {Game[]} games - List of all games matching focus and timespan config options.
      */
     games: [],
+
+    /**
+     * @member {Series[]} playoffSeries - List of all current playoff series.
+     */
+    playoffSeries: [],
+
     /**
      * @member {SeasonDetails} season - Current season details e.g. year and mode.
      */
@@ -83,6 +91,7 @@ Module.register('MMM-NHL', {
      * @property {number} daysAhead - Amount of days a match should be displayed before it starts.
      * @property {boolean} showNames - Flag to show team names.
      * @property {boolean} showLogos - Flag to show club logos.
+     * @property {boolean} showPlayoffSeries - Flag to show playoff series status during playoffs.
      * @property {boolean} rollOverGames - Flag to show today's games and previous/next day based on game status.
      */
     defaults: {
@@ -96,6 +105,7 @@ Module.register('MMM-NHL', {
         daysAhead: 7,
         showNames: true,
         showLogos: true,
+        showPlayoffSeries: true,
         rollOver: false
     },
 
@@ -149,6 +159,7 @@ Module.register('MMM-NHL', {
             modes: this.modes,
             season: this.season,
             games: this.games,
+            playoffSeries: this.playoffSeries,
             rotateIndex: this.rotateIndex,
             maxGames: Math.min(this.games.length, this.rotateIndex + this.config.matches),
             config: this.config
@@ -171,7 +182,7 @@ Module.register('MMM-NHL', {
             this.config.daysAhead = 1;
         }
 
-        this.sendSocketNotification('CONFIG', {config: this.config});
+        this.sendSocketNotification('CONFIG', { config: this.config });
     },
 
     /**
@@ -188,6 +199,9 @@ Module.register('MMM-NHL', {
             this.games = payload.games;
             this.season = payload.season;
             this.setRotateInterval();
+        } else if (notification === 'PLAYOFFS') {
+            this.playoffSeries = payload;
+            this.updateDom(300);
         }
     },